requirements.txt
pytesseract
opencv-python
streamlit
numpy
pillow
